pyproject.toml
[tool.poetry]
name = "paper-scout"
version = "0.1.0"
description = ""
authors = ["Abhay Shukla <abhayshuklaistheone@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mcp = {extras = ["cli"], version = "^1.9.2"}
litellm = "^1.72.0"
python-dotenv = "^1.1.0"
pymupdf = "^1.26.0"
fastmcp = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
